refactor: 코드 함수 단위로 분리
