+++ src/@lib/hooks/index.ts
@@ -2,3 +2,6 @@ export * from "./useDeepMemo";
 export * from "./useMemo";
 export * from "./useCallback";
 export * from "./useRef";
+export * from "./useThemeContext";
+export * from "./useUserContext";
+export * from "./useNotificationContext";
