+++ src/@lib/hooks/useMemo.ts
@@ -5,7 +5,7 @@ import { useRef } from "./useRef";
 export function useMemo<T>(
   factory: () => T,
   _deps: DependencyList,
-  _equals = shallowEquals,
+  _equals = shallowEquals
 ): T {
   const valueRef = useRef<T | undefined>(undefined);
   const depsRef = useRef<DependencyList | undefined>(undefined);
