+++ src/@lib/hooks/useNotificationContext.ts
@@ -1,0 +1,12 @@
+import { useContext } from "react";
+import { NotificationContext } from "../../context";
+
+export const useNotificationContext = () => {
+  const context = useContext(NotificationContext);
+  if (context === undefined) {
+    throw new Error(
+      "useNotificationContext must be used within a NotificationProvider"
+    );
+  }
+  return context;
+};
