+++ src/@lib/hooks/useUserContext.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { UserContext } from "../../context";
+
+export const useUserContext = () => {
+  const context = useContext(UserContext);
+  if (context === undefined) {
+    throw new Error("useUserContext must be used within a UserProvider");
+  }
+  return context;
+};
