+++ src/context/index.ts
@@ -1,0 +1,3 @@
+export * from "./themeContext";
+export * from "./userContext";
+export * from "./notificationContext";
