+++ src/context/userContext.ts
@@ -1,0 +1,12 @@
+import { createContext } from "react";
+import { User } from "../types";
+
+interface UserContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+export const UserContext = createContext<UserContextType | undefined>(
+  undefined
+);
